Split app layout test definition files
