+++ test/definitions/index.ts
@@ -7,9 +7,35 @@ import { TestSuite } from './types';
 import actionCard from './visual/action-card';
 import alert from './visual/alert';
 import appLayout from './visual/app-layout';
+import appLayoutContentPaddings from './visual/app-layout-content-paddings';
+import appLayoutDrawers from './visual/app-layout-drawers';
+import appLayoutFlashbar from './visual/app-layout-flashbar';
+import appLayoutHeader from './visual/app-layout-header';
+import appLayoutMulti from './visual/app-layout-multi';
+import appLayoutResponsive from './visual/app-layout-responsive';
+import appLayoutStickyTableHeaderSplitPanel from './visual/app-layout-sticky-table-header-split-panel';
+import appLayoutToolbar from './visual/app-layout-toolbar';
+import appLayoutZIndex from './visual/app-layout-z-index';
 import areaChart from './visual/area-chart';
 import attributeEditor from './visual/attribute-editor';
 import autosuggest from './visual/autosuggest';
 import badge from './visual/badge';
 
-export const allSuites: TestSuite[] = [actionCard, alert, appLayout, areaChart, attributeEditor, autosuggest, badge];
+export const allSuites: TestSuite[] = [
+  actionCard,
+  alert,
+  appLayout,
+  appLayoutContentPaddings,
+  appLayoutDrawers,
+  appLayoutFlashbar,
+  appLayoutHeader,
+  appLayoutMulti,
+  appLayoutResponsive,
+  appLayoutStickyTableHeaderSplitPanel,
+  appLayoutToolbar,
+  appLayoutZIndex,
+  areaChart,
+  attributeEditor,
+  autosuggest,
+  badge,
+];
