+++ test/visual/app-layout-multi.test.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import { runTestSuites } from '../definitions/utils';
-import suite from '../definitions/visual/app-layout';
+import suite from '../definitions/visual/app-layout-multi';
 
 runTestSuites([suite]);
